Move storage package
