+++ Corona-Warn-App/src/main/java/de/rki/coronawarnapp/presencetracing/storage/retention/TraceLocationDbCleanUpPeriodicWorker.kt
@@ -1,4 +1,4 @@
-package de.rki.coronawarnapp.eventregistration.storage.retention
+package de.rki.coronawarnapp.presencetracing.storage.retention
 
 import android.content.Context
 import androidx.work.CoroutineWorker
